+++ Automate/Framework/AutomateConstants.cs
@@ -1,7 +1,11 @@
+//
+// Note: this file is shared between the Automate and Chests Anywhere projects.
+//
+
 namespace Pathoschild.Stardew.Automate.Framework
 {
     /// <summary>Provide constants for Automate logic.</summary>
-    internal static class ModConstants
+    internal static class AutomateConstants
     {
         /// <summary>The context tag which indicates an item that can be treated as a storage container.</summary>
         public const string StorageTag = "automate_storage";