+++ Automate/ModEntry.cs
@@ -152,12 +152,12 @@ internal class ModEntry : Mod
             {
                 IDictionary<string, BigCraftableData> assetData = asset.AsDictionary<string, BigCraftableData>().Data;
 
-                foreach (string itemId in ModConstants.GetDefaultChestItemIds())
+                foreach (string itemId in AutomateConstants.GetDefaultChestItemIds())
                 {
                     if (assetData.TryGetValue(itemId, out BigCraftableData? entry))
                     {
                         entry.ContextTags ??= [];
-                        entry.ContextTags.Add(ModConstants.StorageTag);
+                        entry.ContextTags.Add(AutomateConstants.StorageTag);
                     }
                 }
             });