+++ ChestsAnywhere/docs/release-notes.md
@@ -1,7 +1,9 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
-## Upcoming release
+## 1.28.3
+Released 27 July 2025 for SMAPI 4.1.10 or later.
+
 * Updated chest options to support Automate 2.4.0+ configurable chest types.
 
 ## 1.28.2